Project

+++ ZubrilkaEnglish(android)/app/src/main/java/com/example/zubrilkaenglish/repositories/Repository.kt
@@ -3,6 +3,8 @@ package com.example.zubrilkaenglish.repositories
 import com.example.zubrilkaenglish.models.Word
 import com.example.zubrilkaenglish.repositories.retrofit.RetrofitService
 import com.example.zubrilkaenglish.repositories.room.RoomService
+import org.threeten.bp.ZonedDateTime
+import org.threeten.bp.format.DateTimeFormatter
 
 class Repository {
     private val retrofitService=RetrofitService()
@@ -13,25 +15,51 @@ class Repository {
      */
     suspend fun getAllWords():List<Word>{
 
-        checkUpdateAtOnServer()
+        if(checkUpdateAtOnServer()){
+            val list=retrofitService.getAllWords()
+            if (list != null) {
+                println("получили список с инета. Его размер = "+list.size)
+                roomService.deleteAllWords()
+                println("удалили старый список из БД")
+                roomService.insertListWords(list)
+                println("положили список в БД")
 
-        val list=retrofitService.getAllWords()
-        if (list != null) {
-            println("получили список с инета. Его размер = "+list.size)
-            roomService.deleteAllWords()
-            println("удалили старый список из БД")
-            roomService.insertListWords(list)
-            println("положили список в БД")
-        }else{
-            println("список с сервера был null")
+                val newUpdateAt=retrofitService.getUpdateAt()
+                roomService.insertNewUpdatedAt(newUpdateAt)
+
+            }else{
+                println("список с сервера был null")
+            }
         }
         return roomService.getAllWords()
     }
+
+    /**
+     * если на сервере имеется более свежая версия списков данных, метод вернет true или если в базе нет сведений о дате последней версии вернет true
+     */
     private suspend fun checkUpdateAtOnServer():Boolean{
         val serverUpAt: String? =retrofitService.getUpdateAt()
         println(serverUpAt)
-        val roomUpAt:String=roomService.getUpdateAt()
+        val roomUpAt:String?=roomService.getUpdatedAt()
         println(roomUpAt)
-        return true
+
+        if (serverUpAt==null){
+            println("serverUpAt is null")
+            return false
+        }else if(roomUpAt==null){
+            println("roomUpAt is null")
+            return true
+        }else {
+            val parsedServUpAt= ZonedDateTime.parse(serverUpAt, DateTimeFormatter.ISO_ZONED_DATE_TIME)
+            val parsedRoomUpAt= ZonedDateTime.parse(roomUpAt, DateTimeFormatter.ISO_ZONED_DATE_TIME)
+
+            if(parsedServUpAt.isAfter(parsedRoomUpAt)){
+                println("parsedServUpAt is after parsedRoomUpAt")
+                return true
+            }else {
+                println("parsedServUpAt is before or is equals parsedRoomUpAt")
+                return false
+            }
+        }
     }
 }
